feat: mvp header and banner
