+++ apps/zotmeal/src/components/header/banner.tsx
@@ -1,10 +1,10 @@
-import { useCallback, useMemo, useRef, useState } from "react";
+import { useCallback, useRef } from "react";
 import {
   Button,
-  Dimensions,
   Image,
-  type ImageSourcePropType,
+  Text,
   View,
+  type ImageSourcePropType,
 } from "react-native";
 import Carousel, {
   type ICarouselInstance,
@@ -13,6 +13,7 @@ import { IS_WEB, type Restaurant } from "../../lib/constants";
 import type { CarouselRenderItemInfo } from "react-native-reanimated-carousel/lib/typescript/types";
 import { useRestaurantStore } from "../../stores/restaurant";
 import { useThemeStore } from "../../stores/theme";
+import { AntDesign, Feather, FontAwesome } from '@expo/vector-icons';
 
 interface BannerInfo {
   name: Restaurant;
@@ -33,7 +34,7 @@ const banners: BannerInfo[] = [
 export function Banner() {
   const setRestaurant = useRestaurantStore((store) => store.setRestaurant);
 
-  const width = useThemeStore(store => store.width)
+  const width = useThemeStore((store) => store.width);
 
   const ref = useRef<ICarouselInstance | null>(null);
 
@@ -67,17 +68,33 @@ export function Banner() {
   );
 
   return (
-    <View className="p-2 justify-center items-center" style={{ width }}>
-      <Carousel
-        width={width}
-        height={200}
-        onProgressChange={onProgressChange}
-        data={banners}
-        ref={ref}
-        renderItem={renderItem}
-      />
+    <View className="relative">
+      <View className="w-full h-12 p-2 bg-black/70 absolute z-10 flex-row justify-between items-center">
+        <View className="flex flex-row items-center gap-2">
+          <View className="w-4 h-4 bg-red-400 rounded-full" />
+          <Text className="text-white text-lg font-bold">Closed Now</Text>
+        </View>
+
+        <View className="flex-row gap-4">
+          <AntDesign name="calendar" size={24} color="white" />
+          <FontAwesome name="dollar" size={24} color="white" />
+          <Feather name="map-pin" size={24} color="white" />
+        </View>
+      </View>
+
+      <View className="justify-center items-center" style={{ width }}>
+        <Carousel
+          width={width}
+          height={200}
+          onProgressChange={onProgressChange}
+          data={banners}
+          ref={ref}
+          renderItem={renderItem}
+        />
+      </View>
+
       {IS_WEB && (
-        <View className="flex-row self-end gap-2 my-1">
+        <View className="p-2 my-2 flex-row self-end gap-2">
           {banners.map((banner, index) => {
             return (
               <View key={banner.source.toString()}>
